Split scaffold, smoke, and form helpers
