+++ .sampo/changesets/issue-420-runtime-boundary-batch.md
@@ -1,0 +1,6 @@
+---
+npm/wp-typia: patch
+npm/@wp-typia/project-tools: patch
+---
+
+Split generated project smoke assertions, scaffold runtime helpers, and first-party form helpers into focused modules while keeping the existing CLI and TUI behavior stable.
